+++ backend/botler/api/__init__.py
@@ -17,6 +17,7 @@ from .backup import router as backup_router  # noqa: E402
 from .notifications import router as notifications_router  # noqa: E402
 from .environment import router as environment_router  # noqa: E402
 from .auth import router as auth_router  # noqa: E402
+from .labels import router as labels_router  # noqa: E402
 
 router.include_router(repos_router)
 router.include_router(tasks_router)
@@ -25,3 +26,4 @@ router.include_router(backup_router)
 router.include_router(notifications_router)
 router.include_router(environment_router)
 router.include_router(auth_router)
+router.include_router(labels_router)